Rename dataset module -> metaframe
