+++ docs/source/usage/generating_metaframes.rst
@@ -1,7 +1,7 @@
 Generating MetaFrames
 =====================
 
-One of the main features of metasyn is to create a :obj:`MetaFrame <metasyn.dataset.MetaFrame>`, an object which captures the essential aspects of the dataset, including variable names, types, data types, the percentage of missing values, and distribution attributes. :obj:`MetaFrame <metasyn.dataset.MetaFrame>` objects essentially capture all the information needed to generate a synthetic dataset that aligns with the original dataset.
+One of the main features of metasyn is to create a :obj:`MetaFrame <metasyn.metaframe.MetaFrame>`, an object which captures the essential aspects of the dataset, including variable names, types, data types, the percentage of missing values, and distribution attributes. :obj:`MetaFrame <metasyn.metaframe.MetaFrame>` objects essentially capture all the information needed to generate a synthetic dataset that aligns with the original dataset.
 
 .. image:: /images/pipeline_estimation_simple.png
    :alt: MetaFrame Generation Flow
@@ -15,13 +15,13 @@ One of the main features of metasyn is to create a :obj:`MetaFrame <metasyn.data
 Basics
 ------
 
-Metasyn can generate metadata from any given dataset (provided as Polars or Pandas DataFrame), using the :meth:`metasyn.MetaFrame.fit_dataframe(df) <metasyn.dataset.MetaFrame.fit_dataframe>` classmethod.
+Metasyn can generate metadata from any given dataset (provided as Polars or Pandas DataFrame), using the :meth:`metasyn.MetaFrame.fit_dataframe(df) <metasyn.metaframe.MetaFrame.fit_dataframe>` classmethod.
 
 .. image:: /images/pipeline_estimation_code.png
    :alt: MetaFrame Generation With Code Snippet
    :align: center
 
-This function requires a :obj:`DataFrame` to be specified as parameter. The following code returns a :obj:`MetaFrame<metasyn.dataset.MetaFrame>` object named :obj:`mf`, based on a DataFrame named :obj:`df`.
+This function requires a :obj:`DataFrame` to be specified as parameter. The following code returns a :obj:`MetaFrame<metasyn.metaframe.MetaFrame>` object named :obj:`mf`, based on a DataFrame named :obj:`df`.
 
 .. code-block:: python
     
@@ -29,15 +29,15 @@ This function requires a :obj:`DataFrame` to be specified as parameter. The foll
 
 .. admonition:: Note on Pandas and Polars DataFrames
 
-    Internally, metasyn uses Polars (instead of Pandas) mainly because typing and the handling of non-existing data is more consistent. It is possible to supply a Pandas DataFrame instead of a Polars DataFrame to ``MetaDataset.from_dataframe``. However, this uses the automatic Polars conversion functionality, which for some edge cases result in problems. Therefore, we advise users to create Polars DataFrames. The resulting synthetic dataset is always a Polars DataFrame, but this can be easily converted back to a Pandas DataFrame by using ``df_pandas = df_polars.to_pandas()``.
+    Internally, metasyn uses Polars (instead of Pandas) mainly because typing and the handling of non-existing data is more consistent. It is possible to supply a Pandas DataFrame instead of a Polars DataFrame to the ``MetaFrame.from_dataframe`` method. However, this uses the automatic Polars conversion functionality, which for some edge cases result in problems. Therefore, we advise users to create Polars DataFrames. The resulting synthetic dataset is always a Polars DataFrame, but this can be easily converted back to a Pandas DataFrame by using ``df_pandas = df_polars.to_pandas()``.
 
 .. _OptionalParams:
 
 Optional Parameters
 ----------------------
-The :meth:`metasyn.MetaFrame.fit_dataframe() <metasyn.dataset.MetaFrame.fit_dataframe>` class method allows you to have more control over how your synthetic dataset is generated with additional (optional) parameters:
+The :meth:`metasyn.MetaFrame.fit_dataframe() <metasyn.metaframe.MetaFrame.fit_dataframe>` class method allows you to have more control over how your synthetic dataset is generated with additional (optional) parameters:
     
-Besides the required `df` parameter, :meth:`metasyn.MetaFrame.fit_dataframe() <metasyn.dataset.MetaFrame.fit_dataframe>` accepts three parameters: ``spec``, ``dist_providers`` and ``privacy``.
+Besides the required `df` parameter, :meth:`metasyn.MetaFrame.fit_dataframe() <metasyn.metaframe.MetaFrame.fit_dataframe>` accepts three parameters: ``spec``, ``dist_providers`` and ``privacy``.
 
 Let's take a look at each optional parameter individually:
