+++ docs/source/usage/generating_synthetic_data.rst
@@ -1,7 +1,7 @@
 Generating Synthetic Data
 =========================
 
-Metasyn can **generate synthetic data** from any :obj:`MetaFrame <metasyn.dataset.MetaFrame>` object.
+Metasyn can **generate synthetic data** from any :obj:`MetaFrame <metasyn.metaframe.MetaFrame>` object.
 
 .. image:: /images/pipeline_generation_simple.png
    :alt: Synthetic Data Generation
@@ -13,21 +13,21 @@ The generated data does **not** aim to preserve the relationships between variab
 
 .. warning:: 
 
-   Before synthetic data can be generated, a :obj:`MetaFrame <metasyn.dataset.MetaFrame>` object must be :doc:`created </usage/generating_metaframes>` or :doc:`loaded </usage/exporting_metaframes>`.
+   Before synthetic data can be generated, a :obj:`MetaFrame <metasyn.metaframe.MetaFrame>` object must be :doc:`created </usage/generating_metaframes>` or :doc:`loaded </usage/exporting_metaframes>`.
 
-To generate a synthetic dataset, simply call the :meth:`MetaFrame.synthesize(n) <metasyn.dataset.MetaFrame.synthesize>` method on a :obj:`MetaFrame <metasyn.dataset.MetaFrame>` object. This method takes an integer parameter `n` which represents the number of rows of data that should be generated. This parameter *must* be specified when calling the method. 
+To generate a synthetic dataset, simply call the :meth:`MetaFrame.synthesize(n) <metasyn.metaframe.MetaFrame.synthesize>` method on a :obj:`MetaFrame <metasyn.metaframe.MetaFrame>` object. This method takes an integer parameter `n` which represents the number of rows of data that should be generated. This parameter *must* be specified when calling the method. 
 
 .. image:: /images/pipeline_generation_code.png
    :alt: Synthetic Data Generation With Code Snippet
    :align: center
 
-The following code generates 5 rows of data based on a :obj:`MetaFrame <metasyn.dataset.MetaFrame>` named ``mf``.
+The following code generates 5 rows of data based on a :obj:`MetaFrame <metasyn.metaframe.MetaFrame>` named ``mf``.
 
 .. code-block:: python
    
    mf.synthesize(5)
 
-Upon succesful execution of the :meth:`MetaFrame.synthesize(n)<metasyn.dataset.MetaFrame.synthesize>` method, a `Polars DataFrame <https://pola-rs.github.io/polars/py-polars/html/reference/dataframe/index.html>`_ will be returned.
+Upon succesful execution of the :meth:`MetaFrame.synthesize(n)<metasyn.metaframe.MetaFrame.synthesize>` method, a `Polars DataFrame <https://pola-rs.github.io/polars/py-polars/html/reference/dataframe/index.html>`_ will be returned.
