+++ metasyn/__init__.py
@@ -1,4 +1,4 @@
-"""MetaSynth: a package for creating synthetic datasets.
+"""Metasyn: a package for creating synthetic datasets.
 
 One part concerns the creation of the statistical metadata from the
 original data, while the other part creates a synthetic dataset from the
@@ -7,7 +7,7 @@ metadata.
 
 from importlib.metadata import version
 
-from metasyn.dataset import MetaFrame
+from metasyn.metaframe import MetaFrame
 from metasyn.var import MetaVar
 from metasyn.demo.dataset import demo_file
 from metasyn.distribution.base import metadist
